refactor: 서버 분리를 대응하는 네트워크 레이어 개선
