+++ AnnualLeaveCalculator/AnnualLeaveCalculator/App/Info.plist
@@ -2,8 +2,6 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
-	<key>ITSAppUsesNonExemptEncryption</key>
-	<false/>
 	<key>BASE_URL</key>
 	<string>$(BASE_URL)</string>
 	<key>UIAppFonts</key>
